backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
supabase>=2.3.0
httpx>=0.26.0
google-genai>=1.0.0
google-genai>=1.0.0
pywebpush>=1.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
